+++ frontend/src/components/Upload/CustomInput.tsx
@@ -2,7 +2,10 @@
 import { IInputProps } from '@jeffreyca/react-dropzone-uploader';
 import { getDroppedOrSelectedFiles } from 'html5-file-selector';
 import * as React from 'react';
-import { CloudUpload } from 'react-bootstrap-icons';
+import { OverlayTrigger, Tooltip } from 'react-bootstrap';
+import { CloudUpload, InfoCircle } from 'react-bootstrap-icons';
+import { OverlayInjectedProps } from 'react-bootstrap/esm/Overlay';
+import { ALLOWED_EXTENSIONS } from '../../Constants';
 
 /**
  * Custom file input component for the dropzone uploader.
@@ -21,10 +24,27 @@ const CustomInput = ({ accept, onFiles, files, disabled }: IInputProps): JSX.Ele
     });
   };
 
+  const supportedFormatsTooltip = (props: OverlayInjectedProps) => {
+    const text = 'Supported: ' + ALLOWED_EXTENSIONS.sort().join(', ');
+    console.log(text);
+
+    return (
+      <Tooltip id="status-tooltip" {...props}>
+        {text}
+      </Tooltip>
+    );
+  };
+
   return files.length > 0 ? null : (
     <div className="text-center p-3">
       <CloudUpload color="grey" size={70} />
-      <p>Drag and drop an audio file (.mp3, .flac, .wav)</p>
+      <p className="d-flex align-items-center justify-content-start">
+        Drag and drop an audio file. {'  '}
+        <OverlayTrigger placement="right" delay={{ show: 100, hide: 100 }} overlay={supportedFormatsTooltip}>
+          <InfoCircle className="ml-1" size={14} style={{ cursor: 'pointer' }} />
+        </OverlayTrigger>
+      </p>
+
       <label className={buttonClass}>
         {text}
         <input